pyproject.toml
[tool.poetry]
name = "serviceemail"
version = "0.1.0"
description = ""
authors = ["matheucdd, leomitas, alves-leandro, PauloHenriqueTCastro"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
